Replace executor with blocking queue and a thread
Previously, the ProcessorImportService was implemented as a single
thread executor, and callables were added to it for every log that
needed imported. Now, it is a simple blocking queue and a thread that
will loop over that queue as needed. Yes its basically the same thing
but this was allows me to better control when to stop the importing
loop.