Create Log and ImportInfo at same time
Previously:
The import operation creates the ImportInfo entry first to ensure that
it is a legitimate jmeter jtl file. Once the import succeeds, the log
entry is created.

This made editing the log details difficult. The idea is that the user
uploads a file and can change details, like the name and description,
while the file is getting processed. However, if the user edits the
ImportInfo while it is getting prcessed, and it then successfully turns
into a Log before the user could save the new details, the changes
could be lost.

Now:
The import operation creates a Log and an ImportInfo entry. The Log
has a status field. If there is an error with the import, it can
be indicated with the status field.