Issue #2 Fix review log-applicability calculation interference
Previously, it was possible to submit a review with an invalid label
selector. When a new log is added, all the reviews are recalculated to see
if that log applies to them. If one of those reviews had an invalid label
selector, a LabelSelectorSyntaxException was thrown, and would prevent the
rest of the reviews from getting recalculated.

This fix does the following:
- If a new review is created and the label selector is invalid, the
  selector is tossed out.
- If a review is edited and the new label selector is invalid, the old
  selector is used.
- If any reviews still happen to have an invalid label selector, an error
  is logged and the next review is calculated.