Add way to get multiple labels of code counts in single call
The code_count table is different from the rest of the stats tables:
The other stats tables have an aggregate table and a timeseries table,
and corresponding getters in StatsService; code_count combines the
aggregate and timeseries data into a single table, and so the getters
must be able to let the client retrieve the aggregate vs. the timeseries
tables.