Issue #26 HTTP API listing logs from label selector
This adds the ability to list logs via the HTTP API with a label
selector. Example:

curl -H 'Accept: application/json' "$BASE_URL/logs/?labelSelector=example"

This will return a list of logs that match the label "example", which
would include the ids, which can be used to retrieve the logs.