Fix NPE on empty Status Message
When StatsBuilder builds stats from a JTL file, and the samples are
sorted, it is possible for comparisons to compare by Status Message if
the lines are sufficiently identical. When this happens, and the Status
Message is empty, the converter treats the entry as null, which triggers
an NPE. This is now fixed.

Additionally, the jooq maven plugin required Java 17 in order to run.
This commit rolls back to the most recent jooq libs and plugin that
doesn't require Java 17 to build and run.