Fix #33 Use pure.css
This switches to using pure.css, switching off of bootstrap and jquery.
Also, feather, glyphicons, and chart.js are removed.

Additionally, on pages that have tabbed content, switching tabs will now
update the url hash, which allows those URLs to be visited, and would
then automatically open up the relevant tab as a result. (Previously,
clicking a tab would not result in updating the hash, and if the hash
was added to the URL anyway, the tab would not switch.)

Remove bootstrap, jquery, feature, glyphicons, chart.js

These are now unused.